Move to companion object
