+++ core/jvm/src/main/scala/zio/PlatformSpecific.scala
@@ -23,7 +23,7 @@ import zio.random.Random
 import zio.scheduler.Scheduler
 import zio.system.System
 
-private[zio] trait PlatformSpecific {
+private[zio] trait PlatformSpecific extends JavaSpecific {
   type ZEnv = Clock with Console with System with Random with Scheduler with Blocking
 
   object ZEnv {
